# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Price Tag API",
    "description": "LLM-friendly REST API for creating, managing, and generating professional price tags with PDF export. This API is optimized for function calling by AI assistants and LLMs.",
    "version": "1.0.0",
    "contact": {
      "name": "Price Tag API Support",
      "url": "https://t.me/PriceTagPrinterBot"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://your-domain.com/api",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/price-tags/{id}": {
      "get": {
        "operationId": "getPriceTag",
        "summary": "Get a specific price tag by ID",
        "description": "Retrieve detailed information about a single price tag including all pricing tiers and design settings.",
        "tags": ["Price Tags"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Price tag ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Price tag details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SinglePriceTagResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      },
      "put": {
        "operationId": "updatePriceTag",
        "summary": "Update a specific price tag",
        "description": "Update any field of an existing price tag. Discount prices are automatically recalculated when price is updated.",
        "tags": ["Price Tags"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Price tag ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePriceTagRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Price tag updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SinglePriceTagResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      },
      "delete": {
        "operationId": "deletePriceTag",
        "summary": "Delete a specific price tag",
        "description": "Permanently delete a price tag. This action cannot be undone.",
        "tags": ["Price Tags"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Price tag ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Price tag deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "PriceTag": {
        "type": "object",
        "required": ["id", "data", "price", "discountPrice"],
        "properties": {
          "id": {
            "type": "integer",
            "description": "Unique identifier for the price tag"
          },
          "data": {
            "oneOf": [
              {"type": "string"},
              {"type": "number"}
            ],
            "description": "Product name or identifier"
          },
          "price": {
            "type": "integer",
            "minimum": 1,
            "description": "Main price in smallest currency unit (e.g., cents)"
          },
          "discountPrice": {
            "type": "integer",
            "minimum": 0,
            "description": "Calculated discount price in smallest currency unit"
          },
          "designType": {
            "type": "string",
            "enum": ["default", "new", "sale", "white", "black", "sunset", "ocean", "forest", "royal", "vintage", "neon", "monochrome", "silver", "charcoal", "paper", "ink", "snow"],
            "description": "Theme type for visual design"
          },
          "hasDiscount": {
            "type": "boolean",
            "description": "Whether discount pricing should be displayed"
          },
          "priceFor2": {
            "type": "integer",
            "minimum": 1,
            "description": "Special price when buying 2 items"
          },
          "priceFrom3": {
            "type": "integer",
            "minimum": 1,
            "description": "Special price when buying 3 or more items"
          }
        }
      },
...
      "UpdatePriceTagRequest": {
        "type": "object",
        "properties": {
          "data": {
            "oneOf": [
              {"type": "string", "minLength": 1},
              {"type": "number"}
            ],
            "description": "Product name or identifier"
          },
          "price": {
            "type": "integer",
            "minimum": 1,
            "description": "Main price in smallest currency unit"
          },
          "designType": {
            "type": "string",
            "enum": ["default", "new", "sale", "white", "black", "sunset", "ocean", "forest", "royal", "vintage", "neon", "monochrome", "silver", "charcoal", "paper", "ink", "snow"],
            "description": "Theme type for visual design"
          },
          "hasDiscount": {
            "type": "boolean",
            "description": "Whether discount pricing should be displayed"
          },
          "priceFor2": {
            "type": "integer",
            "minimum": 1,
            "description": "Special price when buying 2 items"
          },
          "priceFrom3": {
            "type": "integer",
            "minimum": 1,
            "description": "Special price when buying 3 or more items"
          }
        }
      },
...
      "ApiResponse": {
        "type": "object",
        "required": ["success"],
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Whether the operation was successful"
          },
          "data": {
            "description": "Response data (present when success is true)"
          },
          "error": {
            "type": "string",
            "description": "Error message (present when success is false)"
          },
          "message": {
            "type": "string",
            "description": "Optional additional information"
          }
        }
      },
...
      "SinglePriceTagResponse": {
        "allOf": [
          {"$ref": "#/components/schemas/ApiResponse"},
          {
            "type": "object",
            "properties": {
              "data": {"$ref": "#/components/schemas/PriceTag"}
            }
          }
        ]
      },
...
      "DeleteResponse": {
        "allOf": [
          {"$ref": "#/components/schemas/ApiResponse"},
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "object",
                "properties": {
                  "id": {"type": "integer"}
                }
              }
            }
          }
        ]
      }
    },
    "responses": {
      "ValidationError": {
        "description": "Validation error occurred",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiResponse"
            },
            "example": {
              "success": false,
              "error": "Price is required and must be a positive number"
            }
          }
        }
      },
      "NotFoundError": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiResponse"
            },
            "example": {
              "success": false,
              "error": "Price tag not found"
            }
          }
        }
      }
    }
  },
...
}